Split enrichment to dedicated module
