+++ tests/test_core.py
@@ -1,8 +1,6 @@
 from collections import defaultdict
 
-import pytest
-
-from honeyflare import get_raw_file_entries, get_sampled_file_entries, enrich_entry, compile_pattern, __version__
+from honeyflare import get_raw_file_entries, get_sampled_file_entries, compile_pattern, __version__
 
 
 def test_get_raw_file_entries(test_files):
@@ -50,47 +48,3 @@ def test_get_sampled_file_entries(test_files):
     assert 25 < lines_by_status[403] < 75
     assert lines_by_status[404] == 0
     assert lines_by_status[503] == 200
-
-
-def test_enrich_entry():
-    entry = {
-        'ClientRequestURI': "/users/id1337?regularParam=paramValue&emptyParam=",
-        'EdgeEndTimestamp': 1582850070117000000,
-        'EdgeStartTimestamp': 1582850070112000000,
-        'EdgeRequestHost': 'example.com',
-        'OriginResponseTime': 150*1e6,
-    }
-
-    patterns = [compile_pattern(p) for p in [
-        '/other/thing/:thingId',
-        '/users/:userId/pictures',
-        '/users/:userId',
-    ]]
-    enrich_entry(entry, patterns, None)
-
-    assert entry['DurationSeconds'] == 0.005
-    assert entry['DurationMs'] == 5
-    assert entry['OriginResponseTimeSeconds'] == 0.15
-    assert entry['OriginResponseTimeMs'] == 150
-    assert entry['Path'] == '/users/id1337'
-    assert entry['PathShape'] == '/users/:userId'
-    assert entry['Path_userId'] == 'id1337'
-    assert entry['Query'] == 'regularParam=paramValue&emptyParam='
-    assert entry['QueryShape'] == 'emptyParam=?&regularParam=?'
-    assert entry['Query_emptyParam'] == ''
-    assert entry['Query_regularParam'] == 'paramValue'
-    assert entry['UriShape'] == '/users/:userId?emptyParam=?&regularParam=?'
-
-
-@pytest.mark.parametrize('ip,expected_version', [
-    ('1.2.3.4', 4),
-    ('2001:0db8:85a3:0000:0000:8a2e:0370:7334', 6)
-])
-def test_enrich_origin_ip(ip, expected_version):
-    entry = {
-        'OriginIP': ip,
-    }
-
-    enrich_entry(entry, [], None)
-
-    assert entry['OriginIPVersion'] == expected_version
